feat: add dynamic page title to video detail page (#35)
